requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
pymilvus==2.6.9
pydantic-ai==1.67.0
sentence-transformers==5.2.3
python-dotenv==1.2.2
httpx==0.28.1
openai==2.26.0
